requirements.txt
# Requerimientos del proyecto
Django==5.2.7
django-allauth
deep-translator
google-generativeai
# Instalar con: pip install -r requirements.txt